refactor: split memory API surface into a memory-api crate
The cli, tui, and install crates depended on the full memory crate but
only used its HTTP client, domain types, and SQLite migration helpers -
never the lancedb vector store. That coupling put the entire
lancedb/datafusion/arrow graph (495 of the workspace's 573 dependencies)
in the build closure of every API consumer.

Move types, client, entity, and migration into a new memory-api crate
with only lightweight dependencies (serde, reqwest, sea-orm). The memory
service depends on it and re-exports the modules, so its public API and
internal crate:: paths are unchanged. Consumers now depend on memory-api
directly: the cli dependency closure drops from 495 to 348 crates and no
longer waits on the lance graph to compile.
