+++ crates/install/src/migrate.rs
@@ -56,7 +56,7 @@ pub async fn migrate(service: Option<&str>) -> Result<()> {
         print!("  {} {} … ", "→".cyan(), svc.name.green());
 
         let result = match svc.name {
-            "memory" => memory::apply_migrations_for_path(&db_path).await,
+            "memory" => memory_api::apply_migrations_for_path(&db_path).await,
             "notify" => notify::apply_migrations_for_path(&db_path).await,
             "orchestrator" => orchestrator::apply_migrations_for_path(&db_path).await,
             "communicate" => communicate::apply_migrations_for_path(&db_path).await,
@@ -97,7 +97,7 @@ pub async fn migrate_status(service: Option<&str>) -> Result<()> {
         }
 
         let result = match svc.name {
-            "memory" => memory::migration_status_for_path(&db_path).await,
+            "memory" => memory_api::migration_status_for_path(&db_path).await,
             "notify" => notify::migration_status_for_path(&db_path).await,
             "orchestrator" => orchestrator::migration_status_for_path(&db_path).await,
             "communicate" => communicate::migration_status_for_path(&db_path).await,
