+++ crates/memory-api/src/entity/memory_entry.rs
