+++ crates/memory-api/src/entity/mod.rs
