+++ crates/memory-api/src/types/requests.rs
