+++ crates/tui/src/control/app.rs
@@ -2,8 +2,8 @@ use super::stream;
 use crate::config::TuiConfig;
 use agentd_common::config::AgentdConfig;
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
-use memory::client::MemoryClient;
-use memory::types::{Memory, SearchRequest};
+use memory_api::client::MemoryClient;
+use memory_api::types::{Memory, SearchRequest};
 use orchestrator::client::OrchestratorClient;
 use orchestrator::scheduler::types::{DispatchResponse, WorkflowResponse};
 use orchestrator::types::AgentResponse;
